webhook_service/requirements.txt
fastapi
uvicorn
psycopg
redis
kafka-python
python-dotenv
psycopg-pool
psycopg-binary